moved the model back in the model file
