+++ database-mongo/index.js
@@ -13,6 +13,22 @@ db.once('open', function() {
   console.log('mongoose connected successfully');
 });
 
+var overviewSchema = mongoose.Schema({
+  prod_name: String,
+  specs: Array,
+  included: Array,
+  boxContent: Array,
+  descriptionHeader: Array,
+  description: Array,
+  imageHeader: Array,
+  images: Array,
+  shippingDate: String,
+  details: Array,
+  html: String
+});
+
+
+var Overview = mongoose.model('Overview', overviewSchema);
 
 var getOverviewData = (params) => {
   return Overview.find({prod_name: params});
@@ -23,4 +39,4 @@ var getHtmlData = (params) => {
   return overviewMain.find();
 }
 
-module.exports = {getOverviewData, getHtmlData, db};
+module.exports = {getOverviewData, getHtmlData, db, Overview};
